requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
python-dotenv==1.0.0
aiofiles==23.2.1 